Replace useState+useEffect with useSyncExternalStore in useUnitPreference
The previous implementation initialised to "metric" on both server and
client and then corrected to the stored preference in a useEffect, which
caused a visible flash on the first paint for any user who had chosen
US or UK.

useSyncExternalStore reads getSnapshot() synchronously during rendering
so the correct preference is in place before the browser paints.
getServerSnapshot returns "metric" to keep SSR hydration consistent.

A module-level listener set means all hook instances share one store, so
the React Context / UnitPreferenceProvider is no longer necessary and
has been removed.
